requirements.txt
Flask==3.1.2
python-dotenv==1.2.1
google-adk==1.19.0
google-genai==1.52.0
kaggle==1.8.2
pandas==2.3.3
fastmcp==2.13.1
mcp==1.22.0
kagglehub>=0.2.0

